Add Create Entry Mutations & Rename Other -> OtherMedia
