+++ frontend/src/Student/mutations/createOtherMediaEntry.graphql
@@ -1,0 +1,5 @@
+mutation createOtherMediaEntry($input: OtherMediaInput!) {
+  createOtherMedia(input: $input) {
+    id
+  }
+}
