+++ frontend/src/Student/mutations/createVideoEntry.graphql
@@ -1,0 +1,5 @@
+mutation createVideoEntry($input: VideoInput!) {
+  createVideo(input: $input) {
+    id
+  }
+}
